refactor : CarTest -> CarNameTest 로 테스트 클래스 별로 하나의 클래스만 테스트 하도록 변경 / 컴파일 안되던 문제 수정
